% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_area_threshold.R
\name{areaPerParea}
\alias{areaPerParea}
\title{Compute percent area covered by process areas}
\usage{
areaPerParea(process_area, dem)
}
\arguments{
\item{process_area}{RasterLayer of GPP process area}

\item{dem}{A DEM as a RasterLayer}
}
\value{
Area covered by process area
}
\description{
Compute percent area covered by process areas
}
