% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcm_performance.R
\name{errMinBboxLength}
\alias{errMinBboxLength}
\title{PCM Runout Distance Error}
\usage{
errMinBboxLength(obs_poly, pred_raster, dem)
}
\arguments{
\item{obs_poly}{The observed/mapped runout track as a SpatialPolygonsDataFrame}

\item{pred_raster}{A RasterLayer object predicting runout}
}
\value{
A list of error measures
}
\description{
Using min. area bounding boxes to calculate predicted length, it measures
    the relative error, relative difference and error.
}
