% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runout_geometry.R
\name{EuclDist}
\alias{EuclDist}
\title{Calculate Euclidean distance between two points}
\usage{
EuclDist(p1x, p1y, p2x, p2y)
}
\arguments{
\item{p1x}{x coordinate of point 1}

\item{p1y}{y coordinate of point 1}

\item{p2x}{x coordinate of point 2}

\item{p2y}{y coordinate of point 2}
}
\value{
The distance between two points
}
\description{
Calculate Euclidean distance between two points
}
\examples{
EuclDist(2,3,7,8)
}
