% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridsearch_gpp_randomwalk.R
\name{getRndWalkOptParams}
\alias{getRndWalkOptParams}
\title{Get random walk grid search optimal parameters}
\usage{
getRndWalkOptParams(
  workspace = getwd(),
  rwslp_vec,
  rwexp_vec,
  rwper_vec,
  n_train,
  measure = median
)
}
\arguments{
\item{workspace}{The file path where the performance results were saved}

\item{rwslp_vec}{The vector of random walk slope thresholds used in the grid search}

\item{rwexp_vec}{The vector or random walk exponents used in the grid search}

\item{rwper_vec}{The vector or random walk persistence factors used in the grid search}

\item{n_train}{Number of runout polygons processed with grid search}

\item{measure}{A measure (e.g. median, mean) to find optimal parameter across grid search space}

\item{slide_id}{Selects a single runout polygon from slide_plys by row}
}
\value{
A dataframe  with the optimal parameter set and AUROC performance
}
\description{
Get random walk grid search optimal parameters
}
