% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runout_geometry.R
\name{getVertices}
\alias{getVertices}
\title{Get vertices from a polygon}
\usage{
getVertices(x)
}
\arguments{
\item{x}{A SpatialPolygon object}
}
\value{
A matrix with all verticies
}
\description{
Get vertices from a polygon
}
\examples{
slide_plys <- rgdal::readOGR(system.file("extdata/dflow_runout_ply.shp", package="runout.opt"))
getVertices(slide_plys)
}
