% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runout_geometry.R
\name{minbb}
\alias{minbb}
\title{Minimum area bounding rectangles}
\usage{
minbb(pts)
}
\arguments{
\item{pts}{A matrix with x and y coordinates of points}
}
\value{
A list including the bounding box coordinates
}
\description{
Computes a minimum area bounding rectangle for a set of points
}
\examples{
pts=cbind(runif(60),runif(60))
bb=minbb(pts)
plot(bb$box,type="l")
points(pts)
pts2=rotxy(pts,pi/1.6)
bb2=minbb(pts2)
plot(bb2$box,type="l")
points(pts2)
}
