% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcm_gridsearch.R
\name{pcmGetGrid}
\alias{pcmGetGrid}
\title{Get PCM runout distance grid search values}
\usage{
pcmGetGrid(x, performance = "relerr", measure = "median",
  from_save = FALSE)
}
\arguments{
\item{x}{A list of PCM grid search values}

\item{measure}{A measure (e.g. "median", "mean") to find optimal parameter across grid search space}

\item{from_save}{(logical) if TRUE, will load save files form current working directory}

\item{performace}{Performance measure "relerr" relative error}
}
\value{
A matrix of grid search space with summary values
}
\description{
Get PCM runout distance grid search values
}
