% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcm_gridsearch.R
\name{pcmGetOpt}
\alias{pcmGetOpt}
\title{Get PCM runout distance grid search optimal parameters}
\usage{
pcmGetOpt(
  workspace = getwd(),
  pcm_md_vec,
  pcm_mu_vec,
  n_train,
  performance = "relerr",
  measure = "median"
)
}
\arguments{
\item{workspace}{The file path where the performance results were saved}

\item{pcm_mu_vec}{The vector or random walk exponents used in the grid search}

\item{n_train}{Number of runout polygons processed with grid search}

\item{measure}{A measure (e.g. "median", "mean") to find optimal parameter across grid search space}

\item{slide_id}{Selects a single runout polygon from slide_plys by row}

\item{performace}{Performance measure "relerr" relative error}
}
\value{
A dataframe  with the optimal parameter set and performance
}
\description{
Get PCM runout distance grid search optimal parameters
}
