% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcm_spatial_cross_validation.R
\name{pcmPoolSPCV}
\alias{pcmPoolSPCV}
\title{Pool PCM spatial cross-validation results}
\usage{
pcmPoolSPCV(x, plot_freq = FALSE)
}
\arguments{
\item{x}{The (list) result of pcmSPCV()}

\item{plot_freq}{logical. if TRUE, will a create bubble plot of optimal parameter
set frequencies across grid search space}
}
\value{
A data frame summarizing the frequency and performance of each
    optimal parameter sets
}
\description{
Creates a summary of the frequency and performance (relative error) of optimal parameter
    sets across spatial cross-validation repetitions.
}
