% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcm_spatial_cross_validation.R
\name{pcmSPCV}
\alias{pcmSPCV}
\title{Perform (repeated) spatial cross-validation of PCM optimization}
\usage{
pcmSPCV(slide_plys, n_folds, repetitions, pcm_mu_v, pcm_md_v)
}
\arguments{
\item{n_folds}{The number of cross-validated folds (i.e. partitions)}

\item{repetitions}{Number of cross-validation repetitions}

\item{pcm_mu_v}{A vector of PCM model sliding friction coefficients}

\item{pcm_md_v}{A vector of PCM model mass-to-drag ratios (m)}

\item{x}{A SpatialPolygonsDataframe}
}
\value{
A vector containing numeric labels defining each fold
}
\description{
Perform (repeated) spatial cross-validation of PCM optimization
}
