% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcm_performance.R
\name{performancePCM}
\alias{performancePCM}
\title{PCM runout distance performance}
\usage{
performancePCM(
  dem,
  slide_plys,
  source_pnts,
  slide_id = 1,
  rw_slp = 33,
  rw_ex = 3,
  rw_per = 2,
  pcm_mu = 0.3,
  pcm_md = 75,
  buffer_ext = 500,
  buffer_source = 50,
  gpp_iter = 1000,
  predict_threshold = 0.5,
  plot_eval = FALSE,
  return_features = FALSE
)
}
\arguments{
\item{dem}{A DEM as a RasterLayer object}

\item{slide_plys}{Runout tracks as a SpatialPolygonsDataFrame}

\item{source_pnts}{Source points as a SpatialPointsDataFrame}

\item{slide_id}{Selects a single runout polygon from slide_plys by row}

\item{rw_slp}{Random walk slope threshold - below lasteral spreading is modelled}

\item{rw_ex}{Random walk exponent controlling lateral spread}

\item{rw_per}{Random walk persistence factor to weight flow direction consistency}

\item{pcm_mu}{PCM model sliding friction coefficient}

\item{pcm_md}{PCM model mass-to-drag ratio (m)}

\item{buffer_ext}{Defines buffer distance (in meters) around runout polygon
to crop source DEM. This helps to reduce computational time}

\item{buffer_source}{Can define a buffer distance (in meters) to extend source
point to a source area}

\item{gpp_iter}{Number of random walk model iterations}

\item{predict_threshold}{A cutoff value to define what quantile of simulated runout
frequencies is the predicted runout.}

\item{plot_eval}{If TRUE, will plot random walk path and runout polygon}

\item{return_features}{If TRUE, returned list will include GPP input and output
data, in addition to a list of error measures.}
}
\value{
A list of runout distance performance measures.
}
\description{
Computes the error for runout distances simuluated using the random walk
     and PCM model components of the GPP tool in SAGA-GIS.
}
\details{
Runout is either simulated from a single source point or a buffered
     area round the source point.
}
\examples{
## Not run:
# Initalize a saga object
saga <- Rsagacmd::saga_gis()
}
