% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpp_randomwalk_optimization.R,
%   R/performance_gpp_randomwalk.R
\name{performanceRndWalk}
\alias{performanceRndWalk}
\title{Random walk runout path performance}
\usage{
performanceRndWalk(
  dem,
  slide_plys,
  source_pnts,
  slide_id = 2,
  slp = 33,
  ex = 3,
  per = 2,
  gpp_iter = 1000,
  buffer_ext = 500,
  buffer_source = NULL,
  plot_eval = FALSE
)

performanceRndWalk(
  dem,
  slide_plys,
  source_pnts,
  slide_id = 2,
  slp = 33,
  ex = 3,
  per = 2,
  gpp_iter = 1000,
  buffer_ext = 500,
  buffer_source = NULL,
  plot_eval = FALSE
)
}
\arguments{
\item{dem}{A DEM as a RasterLayer object}

\item{slide_plys}{Runout tracks as a SpatialPolygonsDataFrame}

\item{source_pnts}{Source points as a SpatialPointsDataFrame}

\item{slide_id}{Selects a single runout polygon from slide_plys by row}

\item{slp}{Random walk slope threshold - below lasteral spreading is modelled}

\item{ex}{Random walk exponent controlling lateral spread}

\item{per}{Random walk persistence factor to weight flow direction consistency}

\item{gpp_iter}{Number of random walk model iterations}

\item{buffer_ext}{Defines buffer distance (in meters) around runout polygon
to crop source DEM. This helps to reduce computational time}

\item{buffer_source}{Can define a buffer distance (in meters) to extend source
point to a source area}

\item{plot_eval}{If TRUE, will plot random walk path and runout polygon}
}
\value{
the area under the receiver operating characteristic

the area under the receiver operating characteristic
}
\description{
Computes the area under the receiver operating characteristic curve (AUROC) for
     runout paths simuluated using the random walk model component of the
     GPP tool in SAGA-GIS. The AUROC compares a runout polygon to the simulated
     path.

Computes the area under the receiver operating characteristic curve (AUROC) for
     runout paths simuluated using the random walk model component of the
     GPP tool in SAGA-GIS. The AUROC compares a runout polygon to the simulated
     path.
}
\details{
Runout is either simulated from a single source point or a buffered
     area round the source point.

Runout is either simulated from a single source point or a buffered
     area round the source point.
}
\examples{
## Not run:
# Initalize a saga object
saga <- Rsagacmd::saga_gis()

## Not run:
# Initalize a saga object
saga <- Rsagacmd::saga_gis()

}
