% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomwalk_spatial_cross_validation.R
\name{pool_spcvRndWalk}
\alias{pool_spcvRndWalk}
\title{Pool random walk spatial cross-validation results}
\usage{
pool_spcvRndWalk(x)
}
\arguments{
\item{x}{The (list) result of spcvRndWalk()}
}
\value{
A data frame summarizing the frequency and performance of each
    optimal parameter sets
}
\description{
Creates a summary of the frequency and performance (AUROC) of optimal parameter
    sets across spatial cross-validation repetitions.
}
