% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_rescale.R
\name{rasterCdf}
\alias{rasterCdf}
\title{Cumulative Distribution Function for Rasters}
\usage{
rasterCdf(x)
}
\arguments{
\item{x}{A RasterLayer object}
}
\value{
RasterLayer
}
\description{
Compute an empirical cumulative distribution function for a raster.
}
\examples{
r <- raster(system.file("external/test.grd", package="raster"))
r_ecdf <- RasterCdf(r)
plot(r_ecdf)
}
