% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_rescale.R
\name{rasterMinMaxRescale}
\alias{rasterMinMaxRescale}
\title{Min-max normalization (Rescaling)}
\usage{
rasterMinMaxRescale(x)
}
\arguments{
\item{x}{A RasterLayer object}
}
\value{
RasterLayer
}
\description{
Rescale a raster values in the range from 0 to 1 using min-max normalization.
}
