% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_area_threshold.R
\name{rasterThreshold}
\alias{rasterThreshold}
\title{Raster binary reclassification}
\usage{
rasterThreshold(x, value_range = c(0.9, Inf))
}
\arguments{
\item{x}{A RasterLayer}

\item{value_range}{A vector with the min and max value for reclassification}
}
\value{
A RasterLayer with two classes.
}
\description{
Reclassifies a RasterLayer object based on range of values.
}
