% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runout_geometry.R
\name{rotxy}
\alias{rotxy}
\title{Rotate bounding box}
\usage{
rotxy(pts, angle)
}
\arguments{
\item{pts}{A matrix with x and y coordinates of points}

\item{angle}{Angle to rotate}
}
\value{
A list including the bounding box coordinates
}
\description{
Rotate bounding box
}
\examples{
pts=cbind(runif(60),runif(60))
bb=minbb(pts)
plot(bb$box,type="l")
points(pts)
pts2=rotxy(pts,pi/1.6)
bb2=minbb(pts2)
plot(bb2$box,type="l")
points(pts2)
rotxy(pts, 30)
}
