% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runout_geometry.R
\name{runoutGeom}
\alias{runoutGeom}
\title{Runout Geometry}
\usage{
runoutGeom(runout_plys, elev, ID = NULL, plot.bbox = FALSE)
}
\arguments{
\item{runout_plys}{A SpatialPolygonsDataFrame of runout tracks}

\item{elev}{A RasterLayer DEM}

\item{ID}{NOT SURE...}

\item{plot.bbox}{(logical) Plot bounding box associated with width and length calc.}
}
\value{
A SpatialPolygonsDataFrame with corresponding bounding boxes
}
\description{
Computes runnout attributes: width, length, area, max. elevation,
    min. elevation, and using Spatial Polygons and an DEM
}
