% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_area_threshold.R
\name{runoutPareaPredict}
\alias{runoutPareaPredict}
\title{Runout process areas from source area prediction}
\usage{
runoutPareaPredict(source_pred, dem, source_threshold, rw_slp, rw_exp,
  rw_per, pcm_mu, pcm_md, gpp_iter = 1000, file_name = NULL)
}
\arguments{
\item{source_pred}{RasterLayer with probability of being runout source area}

\item{dem}{A DEM as a RasterLayer object}

\item{rw_slp}{Random walk slope threshold - below lateral spreading is modeled}

\item{rw_exp}{Random walk exponent controlling lateral spread}

\item{rw_per}{Random walk persistence factor to weight flow direction consistency}

\item{pcm_mu}{PCM model sliding friction coefficient}

\item{pcm_md}{PCM model mass-to-drag ratio (m)}

\item{gpp_iter}{Model iterations}
}
\value{
A RasterLayer with runout process areas
}
\description{
Thresholds a prediction raster of source areas (probabilities) and computes
corresponding runout process areas.
}
