% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomwalk_gridsearch.R
\name{rwGetOpt_single}
\alias{rwGetOpt_single}
\title{Get random walk runout distance grid search optimal parameters for single event}
\usage{
rwGetOpt_single(x)
}
\arguments{
\item{x}{the results of the grid search}
}
\value{
A dataframe with the optimal parameter set and performance
}
\description{
Get random walk runout distance grid search optimal parameters for single event
}
