% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomwalk_gridsearch.R
\name{rwGridsearch}
\alias{rwGridsearch}
\title{Grid search optimization for random walk runout path}
\usage{
rwGridsearch(dem, slide_plys, slide_src, slide_id = NULL, slp_v, ex_v,
  per_v, gpp_iter = 1000, buffer_ext = 500, buffer_source = NULL,
  save_res = FALSE, plot_eval = FALSE)
}
\arguments{
\item{dem}{A DEM as a RasterLayer object}

\item{slide_plys}{Runout tracks as a SpatialPolygonsDataFrame}

\item{slide_src}{SSource points as a SpatialPointsDataFrame or source areas
as a SpatialPolygonsDataFrame}

\item{slide_id}{Selects a single runout polygon from slide_plys by row}

\item{slp_v}{A vector of random walk slope thresholds - below lateral spreading is modeled}

\item{ex_v}{A vector or random walk exponents controlling lateral spread}

\item{per_v}{A vector or random walk persistence factors to weight flow direction consistency}

\item{gpp_iter}{Number of model iterations}

\item{buffer_ext}{(Optional) Defines buffer distance (in meters) around runout polygon
to crop source DEM. This helps to reduce computational time}

\item{buffer_source}{(Optional) Can define a buffer distance (in meters) to extend source
point to a source area}

\item{save_res}{(logical), if TRUE, will save results in current working directory}

\item{plot_eval}{If TRUE, will plot random walk path and runout polygon}
}
\value{
the area under the receiver operating characteristic
}
\description{
Computes the area under the receiver operating characteristic curve (AUROC) for
     runout paths simuluated using the random walk model component of the
     GPP tool in SAGA-GIS. The AUROC compares a runout polygon to the simulated
     path.
}
\details{
Runout is either simulated from a single source point or a buffered
     area round the source point.
}
