% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomwalk_spatial_cross_validation.R
\name{rwSPCV}
\alias{rwSPCV}
\title{Perform (repeated) spatial cross-validation of random walk optimization}
\usage{
rwSPCV(slide_plys, n_folds, repetitions, rwslp_vec, rwexp_vec, rwper_vec)
}
\arguments{
\item{n_folds}{The number of cross-validated folds (i.e. partitions)}

\item{repetitions}{Number of cross-validation repetitions}

\item{rwslp_vec}{The vector of random walk slope thresholds used in the grid search}

\item{rwexp_vec}{The vector or random walk exponents used in the grid search}

\item{rwper_vec}{The vector or random walk persistence factors used in the grid search}

\item{x}{A SpatialPolygonsDataframe}
}
\value{
A vector containing numeric labels defining each fold
}
\description{
Perform (repeated) spatial cross-validation of random walk optimization
}
