% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runout_geometry.R
\name{minBBoxSpatialPolygons}
\alias{minBBoxSpatialPolygons}
\title{Minimum area bounding box for spatial polgyons}
\usage{
minBBoxSpatialPolygons(x)
}
\arguments{
\item{x}{A SpatialPolygonsDataFrame}
}
\value{
A SpatialPolygonsDataFrame with corresponding bounding boxes
}
\description{
Determines min. area bounding box for a single or set of spatial polygons
}
