% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcm_gridsearch.R
\name{pcmGetOpt}
\alias{pcmGetOpt}
\title{Get PCM runout distance grid search optimal parameters}
\usage{
pcmGetOpt(x, performance = "relerr", measure = "median",
  from_save = FALSE, plot_opt = FALSE)
}
\arguments{
\item{x}{A list of PCM grid search values}

\item{measure}{A measure (e.g. "median", "mean") to find optimal parameter across grid search space}

\item{from_save}{(logical) if TRUE, will load save files form current working directory}

\item{plot_opt}{(logical) if TRUE, will plot performance across grid search space}

\item{slide_id}{Selects a single runout polygon from slide_plys by row}

\item{performace}{Performance measure "relerr" relative error}
}
\value{
A dataframe  with the optimal parameter set and performance
}
\description{
Get PCM runout distance grid search optimal parameters
}
