% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcm_gridsearch.R
\name{pcmGetOpt_single}
\alias{pcmGetOpt_single}
\title{Get PCM runout distance grid search optimal parameters for single event}
\usage{
pcmGetOpt_single(x, performance = "relerr", from_save = FALSE)
}
\arguments{
\item{x}{A list of PCM grid search values}

\item{performace}{Performance measure "relerr" relative error}
}
\value{
A dataframe  with the optimal parameter set and performance
}
\description{
Get PCM runout distance grid search optimal parameters for single event
}
