% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcm_gridsearch.R
\name{pcmGridsearch}
\alias{pcmGridsearch}
\title{Grid search optimization for PCM runout distance}
\usage{
pcmGridsearch(
  dem,
  workspace = getwd(),
  slide_plys,
  source_pnts,
  slide_id,
  rw_slp,
  rw_ex,
  rw_per,
  pcm_mu_v,
  pcm_md_v,
  gpp_iter = 1000,
  buffer_ext = 500,
  buffer_source = NULL,
  predict_threshold = 0.5,
  plot_eval = FALSE
)
}
\arguments{
\item{dem}{A DEM as a RasterLayer object}

\item{workspace}{The file path where to save performance results for each runout polygon}

\item{slide_plys}{Runout tracks as a SpatialPolygonsDataFrame}

\item{source_pnts}{Source points as a SpatialPointsDataFrame}

\item{slide_id}{Selects a single runout polygon from slide_plys by row}

\item{rw_slp}{Random walk slope threshold - below lasteral spreading is modelled}

\item{rw_ex}{Random walk exponent controlling lateral spread}

\item{rw_per}{Random walk persistence factor to weight flow direction consistency}

\item{pcm_mu_v}{A vector of PCM model sliding friction coefficients}

\item{pcm_md_v}{A vector of PCM model mass-to-drag ratios (m)}

\item{gpp_iter}{Number of random walk model iterations}

\item{buffer_ext}{Defines buffer distance (in meters) around runout polygon
to crop source DEM. This helps to reduce computational time}

\item{buffer_source}{Can define a buffer distance (in meters) to extend source
point to a source area}

\item{predict_threshold}{A cutoff value to define what quantile of simulated runout
frequencies is the predicted runout.}

\item{plot_eval}{If TRUE, will plot simulated runout and runout polygon}
}
\value{
A list with performances for each parameter combinations across grid search space.
     The performances measures include relative error, relative difference, error and AUROC.
}
\description{
Computes performance measures for
     runout paths simuluated using the random walk and PCM model components of the
     GPP tool in SAGA-GIS.
}
\details{
Runout is either simulated from a single source point or a buffered
     area round the source point.
}
