% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcm_spatial_cross_validation.R
\name{pcmSPCV}
\alias{pcmSPCV}
\title{Perform (repeated) spatial cross-validation of PCM optimization}
\usage{
pcmSPCV(x, slide_plys, n_folds, repetitions, from_save = FALSE)
}
\arguments{
\item{x}{A SpatialPolygonsDataframe}

\item{n_folds}{The number of folds (i.e. partitions)}

\item{repetitions}{Number of repetitions}

\item{from_save}{(Logical) if TRUE, will load save files from current working directory}
}
\value{
A vector containing numeric labels defining each fold
}
\description{
Perform (repeated) spatial cross-validation of PCM optimization
}
