% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_area_threshold.R
\name{rocParea}
\alias{rocParea}
\title{Compute area under ROC curve for process areas}
\usage{
rocParea(process_area, slide_polys, smp_size = 1000)
}
\arguments{
\item{process_area}{RasterLayer of GPP process area}

\item{slide_polys}{Runout tracks as a SpatialPolygonsDataFrame}

\item{smp_size}{Size of random sample of runount and non-runout cells}
}
\value{
AUROC
}
\description{
Compute area under ROC curve for process areas
}
