% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runout_geometry.R
\name{runoutGeom}
\alias{runoutGeom}
\title{Runout Geometry}
\usage{
runoutGeom(runout_plys, elev, ID = NULL)
}
\arguments{
\item{runout_plys}{A SpatialPolygonsDataFrame of runout tracks}

\item{elev}{A RasterLayer DEM}

\item{ID}{NOT SURE...}
}
\value{
A data frame containing runout geometries
}
\description{
Computes runnout attributes: width, length, area, max.
    min. elevation, and surface area using Spatial Polygons and a DEM
}
