% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomwalk_spatial_cross_validation.R
\name{rwPoolSPCV}
\alias{rwPoolSPCV}
\title{Pool random walk spatial cross-validation results}
\usage{
rwPoolSPCV(x, plot_freq = FALSE)
}
\arguments{
\item{x}{The (list) result of spcvRndWalk()}

\item{plot_freq}{(Logical) if TRUE, will a create bubble plot of optimal parameter
set frequencies across grid search space}
}
\value{
A data frame summarizing the frequency and performance of each
    optimal parameter sets
}
\description{
Creates a summary of the frequency and performance (AUROC) of optimal parameter
    sets across spatial cross-validation repetitions.
}
