% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomwalk_spatial_cross_validation.R
\name{rwSPCV}
\alias{rwSPCV}
\title{Perform (repeated) spatial cross-validation of random walk optimization}
\usage{
rwSPCV(x, slide_plys, n_folds, repetitions, from_save = FALSE)
}
\arguments{
\item{x}{Random walk grid search results (list) from multiple runouts}

\item{slide_plys}{A SpatialPolygonsDataframe}

\item{n_folds}{The number of cross-validated folds (i.e. partitions)}

\item{repetitions}{Number of cross-validation repetitions}

\item{from_save}{(Logical) if TRUE, will load save files from current working directory}
}
\value{
A vector containing numeric labels defining each fold
}
\description{
Perform (repeated) spatial cross-validation of random walk optimization
}
