% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomwalk_spatial_cross_validation.R
\name{spcvFoldsPoly}
\alias{spcvFoldsPoly}
\title{Make spatial partitions (using k-means clustering)}
\usage{
spcvFoldsPoly(x, n_folds = 5, seed_cv = 11082017)
}
\arguments{
\item{x}{A SpatialPolygonsDataframe}

\item{n_folds}{The number of folds (i.e. partitions)}

\item{seed_cv}{The random seed defined to reproduce results}
}
\value{
A vector containing numeric labels defining each fold
}
\description{
Divides a spatial sample of polygons into irregularly shaped spatial partitions
    using k-means clustring. It uses the spatial coordinates defined by the polygon centroids.
}
